# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface User {
  githubId: number;
  username: string;
  avatarUrl: string;
}

export interface AuthMeResponse {
  user: User & (unknown | null);
}

export interface ConfigSummary {
  id: string;
  name: string;
  layout: string;
  createdAt: string;
  updatedAt: string;
}

export type ConfigState = { [key: string]: unknown };

export interface Config {
  id: string;
  userId: number;
  name: string;
  state: ConfigState;
  layout: string;
  createdAt: string;
  updatedAt: string;
}

export interface ConfigListResponse {
  configs: ConfigSummary[];
}

export interface ConfigResponse {
  config: Config;
}

export type CreateConfigRequestState = { [key: string]: unknown };

export interface CreateConfigRequest {
  name?: string;
  state: CreateConfigRequestState;
}

export type UpdateConfigRequestState = { [key: string]: unknown };

export interface UpdateConfigRequest {
  name?: string;
  state?: UpdateConfigRequestState;
}