# lib/api-zod/src/generated/types/authMeResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthMeResponse {
  user: User & (unknown | null);
}