# lib/api-zod/src/generated/types/config.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConfigState } from './configState';

export interface Config {
  id: string;
  userId: number;
  name: string;
  state: ConfigState;
  layout: string;
  createdAt: Date;
  updatedAt: Date;
}