# lib/api-zod/src/generated/types/configListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConfigSummary } from './configSummary';

export interface ConfigListResponse {
  configs: ConfigSummary[];
}