# lib/api-zod/src/generated/types/configResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Config } from './config';

export interface ConfigResponse {
  config: Config;
}