# lib/api-zod/src/generated/types/configSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConfigSummary {
  id: string;
  name: string;
  layout: string;
  createdAt: Date;
  updatedAt: Date;
}