# lib/api-zod/src/generated/types/createConfigRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateConfigRequestState } from './createConfigRequestState';

export interface CreateConfigRequest {
  name?: string;
  state: CreateConfigRequestState;
}