# lib/api-zod/src/generated/types/updateConfigRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateConfigRequestState } from './updateConfigRequestState';

export interface UpdateConfigRequest {
  name?: string;
  state?: UpdateConfigRequestState;
}